agent/requirements.txt
# OTTO LiveKit Agent Dependencies

# Core LiveKit packages
livekit-agents[silero,turn-detector]>=1.0.0
livekit-plugins-google>=0.10.0
livekit-plugins-noise-cancellation>=0.2.0

# HTTP client for API calls
httpx>=0.27.0

# Web search
duckduckgo-search>=6.0.0

# Environment variables
python-dotenv>=1.0.0

# Token Company compression
tokenc>=0.1.0
